Refactor Schnorr account contract
