+++ yarn-project/aztec.js/src/account_impl/index.ts
@@ -1,7 +1,7 @@
 import { AztecAddress, TxContext } from '@aztec/circuits.js';
 import { ExecutionRequest, TxExecutionRequest } from '@aztec/types';
 
-export * from './account_contract.js';
+export * from './single_key_account_contract.js';
 export * from './account_collection.js';
 
 /** Represents an implementation for a user account contract. Knows how to encode and sign a tx for that particular implementation. */
