+++ yarn-project/aztec.js/src/index.ts
@@ -2,7 +2,6 @@ export * from './contract/index.js';
 export * from './contract_deployer/index.js';
 export * from './utils/index.js';
 export * from './aztec_rpc_client/index.js';
-export * from './auth/index.js';
 export * from './account_impl/index.js';
 
 // TODO - only export necessary stuffs
