+++ yarn-project/end-to-end/src/utils.ts
@@ -1,34 +1,33 @@
 import { AztecNodeConfig, AztecNodeService, getConfigEnvVars } from '@aztec/aztec-node';
-import { Fr } from '@aztec/foundation/fields';
-import { DebugLogger, Logger, createDebugLogger } from '@aztec/foundation/log';
+import { AztecRPCServer, createAztecRPCServer } from '@aztec/aztec-rpc';
 import {
   AccountCollection,
-  AccountContract,
   AccountWallet,
   AztecAddress,
   Contract,
   ContractDeployer,
   EthAddress,
-  SchnorrAuthProvider,
+  SingleKeyAccountContract,
   Wallet,
   generatePublicKey,
 } from '@aztec/aztec.js';
 import { CircuitsWasm, getContractDeploymentInfo } from '@aztec/circuits.js';
 import { Schnorr, pedersenPlookupCommitInputs } from '@aztec/circuits.js/barretenberg';
 import { DeployL1Contracts, deployL1Contract, deployL1Contracts } from '@aztec/ethereum';
+import { ContractAbi } from '@aztec/foundation/abi';
 import { toBigIntBE } from '@aztec/foundation/bigint-buffer';
+import { Fr } from '@aztec/foundation/fields';
+import { DebugLogger, Logger, createDebugLogger } from '@aztec/foundation/log';
 import { PortalERC20Abi, PortalERC20Bytecode, TokenPortalAbi, TokenPortalBytecode } from '@aztec/l1-artifacts';
 import { NonNativeTokenContractAbi, SchnorrAccountContractAbi } from '@aztec/noir-contracts/examples';
-import { ContractAbi } from '@aztec/foundation/abi';
-import { AztecRPCServer, createAztecRPCServer } from '@aztec/aztec-rpc';
 import { randomBytes } from 'crypto';
-import { Account, Chain, HttpTransport, PublicClient, WalletClient, getContract } from 'viem';
-import { mnemonicToAccount } from 'viem/accounts';
 import every from 'lodash.every';
 import zipWith from 'lodash.zipwith';
+import { Account, Chain, HttpTransport, PublicClient, WalletClient, getContract } from 'viem';
+import { mnemonicToAccount } from 'viem/accounts';
 
-import { MNEMONIC, localAnvil } from './fixtures.js';
 import { TxStatus } from '@aztec/types';
+import { MNEMONIC, localAnvil } from './fixtures.js';
 
 /**
  * Sets up the environment for the end-to-end tests.
@@ -79,7 +78,6 @@ export async function setup(numberOfAccounts = 1): Promise<{
   const aztecNode = await AztecNodeService.createAndSync(config);
   const aztecRpcServer = await createAztecRPCServer(aztecNode);
   const accountCollection = new AccountCollection();
-  const wasm = await CircuitsWasm.get();
 
   for (let i = 0; i < numberOfAccounts; ++i) {
     // We use the well-known private key and the validating account contract for the first account,
@@ -112,13 +110,11 @@ export async function setup(numberOfAccounts = 1): Promise<{
     }
     accountCollection.registerAccount(
       deploymentData.address,
-      new AccountContract(
+      new SingleKeyAccountContract(
         deploymentData.address,
-        publicKey,
-        new SchnorrAuthProvider(await Schnorr.new(), privateKey),
         deploymentData.partialAddress,
-        SchnorrAccountContractAbi,
-        wasm,
+        privateKey,
+        await Schnorr.new(),
       ),
     );
     logger(`Created account ${deploymentData.address.toString()} with public key ${publicKey.toString()}`);
