+++ yarn-project/noir-contracts/src/contracts/schnorr_account_contract/src/main.nr
@@ -1,5 +1,6 @@
-// Since we support account abstraction user accounts are represented by a contract. This contract is one such example
-// where the user is authenticated with a Schnorr signature.
+// Account contract that uses Schnorr signatures for authentication. The signing key is the same as the
+// encryption key, and as such is not stored in the contract but part of the address preimage, so it can
+// be verified by passing in the partial contract address.
 contract SchnorrAccount {
     use dep::std;
     use dep::aztec::entrypoint;
@@ -15,45 +16,27 @@ contract SchnorrAccount {
     use dep::aztec::oracle::debug_log::debug_log_format;
     use dep::aztec::generators;
 
-    // A function which has to be implemented by every account abstraction contract.
     fn entrypoint(
         inputs: pub PrivateContextInputs,
         payload: pub EntrypointPayload, // contains a set of arguments, selectors, targets and a nonce
-        owner: pub [u8;64], // x and y coordinates concatenated
-        signature: pub [Field;3], // schnorr signature stored in 3 fields
-        partialContractAddress: pub Field,
+        owner: pub [u8;64], // pubkey x and y coordinates concatenated
+        signature: pub [u8;64], // schnorr signature of the payload hash
+        partial_address: pub Field,
     ) -> distinct pub abi::PrivateCircuitPublicInputs {
-        // 75 = ENTRYPOINT_PAYLOAD_SIZE + 64 + 3 + 1
-        let mut args: BoundedVec<Field, 75> = BoundedVec::new(0);
+        // Initialize context
+        // 136 = ENTRYPOINT_PAYLOAD_SIZE(7) + 64 + 64 + 1
+        let mut args: BoundedVec<Field, 136> = BoundedVec::new(0);
         args = args.push_array(payload.serialize());
-        for byte in owner {
-            args = args.push(byte as Field);
-        }
-        args = args.push_array(signature);
-        args = args.push(partialContractAddress);
-        
+        for byte in owner { args = args.push(byte as Field); }
+        for byte in signature { args = args.push(byte as Field); }
+        args = args.push(partial_address);
         let mut context = Context::new(inputs, abi::hash_args(args.storage));
-
-        // Decompress signature from 3 fields to 64 bytes
-        let mut decompressed_signature: [u8;64] = [0;64];
-
-        let signature_part_0 = signature[0].to_be_bytes(32);
-        for i in 0..31 {
-            decompressed_signature[i] = signature_part_0[i + 1];
-        }
-        let signature_part_1 = signature[1].to_be_bytes(32);
-        for i in 0..31 {
-            decompressed_signature[i + 31] = signature_part_1[i + 1];
-        }
-        let signature_part_2 = signature[2].to_be_bytes(32);
-        for i in 0..2 {
-            decompressed_signature[i + 62] = signature_part_2[i + 1];
-        }
 
         // Verify payload signature
         let payload_bytes: [u8; entrypoint::ENTRYPOINT_PAYLOAD_SIZE_IN_BYTES] = payload.to_be_bytes();
         let payload_hash: [u8; 32] = std::hash::sha256(payload_bytes);
 
+        // Convert owner pubkey into fields
         let mut x: Field = 0;
         let mut y: Field = 0;
         let mut mul: Field = 1;
@@ -65,34 +48,17 @@ contract SchnorrAccount {
           mul *= 256;
         }
         
+        // Verify signature of the payload hash
         // TODO: Find out why this signature verification never fails
-        let verification = std::schnorr::verify_signature(x, y, decompressed_signature, payload_hash);
+        let verification = std::schnorr::verify_signature(x, y, signature, payload_hash);
         assert(verification == true);
 
-        let reproduced_address = dep::std::hash::pedersen_with_separator([x, y, partialContractAddress], generators::ContractAddress)[0];
+        // Verify public key against address
+        let reproduced_address = dep::std::hash::pedersen_with_separator([x, y, partial_address], generators::ContractAddress)[0];
         assert(reproduced_address == inputs.call_context.storage_contract_address);
 
-        // Call private functions
-        for i in 0..entrypoint::ACCOUNT_MAX_PRIVATE_CALLS {
-            let target_address = payload.flattened_targets[i];
-            if target_address != 0 {
-                let function_selector = payload.flattened_selectors[i];
-                let args_hash = payload.flattened_args_hashes[i];
-                let (_callStackItem, updated_context) = PrivateCallStackItem::call_with_packed_args(target_address, function_selector, args_hash, context);
-                context = updated_context;
-            }
-        }
-
-        // Call public functions
-        for i in entrypoint::ACCOUNT_MAX_PRIVATE_CALLS..entrypoint::ACCOUNT_MAX_CALLS {
-            let target_address = payload.flattened_targets[i];
-            if target_address != 0 {
-                let function_selector = payload.flattened_selectors[i];
-                let args_hash = payload.flattened_args_hashes[i];
-                let (_callStackItem, updated_context) = PublicCallStackItem::call_with_packed_args(target_address, function_selector, args_hash, context);
-                context = updated_context;
-            }
-        }
+        // Execute calls
+        context = payload.execute_calls(context);
 
         context.finish()
     }
